Generate a list of all passes in markdown.
This uses the pass registry and a clang-libtool to generate documentation for every pass accessible by the pipeline proto or the --passes argument in a markdown file. The documentation consists of the 'short_name' used to specify the pass, the long-name and the comment associated with the pass definition.

Compound passes defined by the pipeline proto have a new field for any comments.

PiperOrigin-RevId: 776275505