Don't apply loop condition to pipelined loops when it is proven true for the first iteration.
Also add several continuation optimizations to yield a clean data flow graph
for nested pipelined loops.

NOTE: Logic enabling narrowing was temporarily removed from generate_fsm.cc.
This is because when the loop body is unguarded by the loop condition,
as when the first iteration is proven to always have a true condition,
nested loops are broken by this logic. This is because the state element
isn't updated after the last iteration / activation of the (inner) loop, since
the condition being used for this is the jump condition of the loop.

The next_value() generation logic will soon be reworked, so for now this has been
temporarily removed.

PiperOrigin-RevId: 778671647